+++ presto/testing/integration_tests/schemas/tpch/part.sql
@@ -6,6 +6,6 @@ CREATE TABLE hive.tpch_test.part (
     p_type VARCHAR NOT NULL,
     p_size INTEGER NOT NULL,
     p_container VARCHAR NOT NULL,
-    p_retailprice DECIMAL(15,2) NOT NULL,
+    p_retailprice DOUBLE NOT NULL,
     p_comment VARCHAR NOT NULL
 ) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
